Correcoes Finais

+++ ControleCarrosHelenaLuz/Models/DTOs/DTOPreco.cs
@@ -4,7 +4,7 @@
     {
         public DateTime ValidadeInicio { get; set; }
         public DateTime ValidadeFim { get; set; }
-        public decimal ValorHora { get; set; }
-        public decimal ValorAdicional { get; set; }
+        public double ValorHora { get; set; }
+        public double ValorAdicional { get; set; }
     }
 }
