+++ ControleCarrosHelenaLuz/Models/Preco.cs
@@ -10,7 +10,7 @@ namespace ControleCarrosHelenaLuz.Models
         public DateTime ValidadeInicio { get; set; }
 
         public DateTime ValidadeFim { get; set; }
-        public decimal ValorHora { get; set; }
-        public decimal ValorAdicional { get; set; }
+        public double ValorHora { get; set; }
+        public double ValorAdicional { get; set; }
     }
 }
